Move `SystemdServiceClass` module to real `SystemdService` class
